Move sprint endpoints to new `sprints` module
